+++ src/main.rs
@@ -6,7 +6,6 @@ mod deploy;
 mod generate_completion;
 mod get_account;
 mod get_auction_info;
-mod get_balance;
 mod get_chainspec;
 mod get_dictionary_item;
 mod get_era_info;
@@ -26,7 +25,6 @@ use once_cell::sync::Lazy;
 
 use casper_client::{cli, rpcs::results::GetChainspecResult, SuccessResponse};
 
-use crate::query_balance::QueryBalance;
 use account_address::AccountAddress;
 use block::{GetBlock, GetBlockTransfers};
 use command::{ClientCommand, Success};
@@ -36,7 +34,6 @@ use deploy::{
 use generate_completion::GenerateCompletion;
 use get_account::GetAccount;
 use get_auction_info::GetAuctionInfo;
-use get_balance::GetBalance;
 use get_chainspec::GetChainspec;
 use get_dictionary_item::GetDictionaryItem;
 use get_era_info::GetEraInfo;
@@ -46,6 +43,7 @@ use get_state_root_hash::GetStateRootHash;
 use get_validator_changes::GetValidatorChanges;
 use keygen::Keygen;
 use list_rpcs::ListRpcs;
+use query_balance::QueryBalance;
 use query_global_state::QueryGlobalState;
 
 const APP_NAME: &str = "Casper client";
@@ -81,7 +79,6 @@ enum DisplayOrder {
     QueryGlobalState,
     QueryBalance,
     GetDictionaryItem,
-    GetBalance,
     GetAccount,
     GetAuctionInfo,
     GetValidatorChanges,
@@ -114,14 +111,14 @@ fn cli() -> Command<'static> {
             DisplayOrder::GetStateRootHash as usize,
         ))
         .subcommand(GetEraInfo::build(DisplayOrder::GetEraInfo as usize))
-        .subcommand(
-            QueryGlobalState::build(DisplayOrder::QueryGlobalState as usize).alias("query-state"),
-        )
+        .subcommand(QueryGlobalState::build(
+            DisplayOrder::QueryGlobalState as usize,
+        ))
         .subcommand(QueryBalance::build(DisplayOrder::QueryBalance as usize))
         .subcommand(GetDictionaryItem::build(
             DisplayOrder::GetDictionaryItem as usize,
         ))
-        .subcommand(GetBalance::build(DisplayOrder::GetBalance as usize))
+        // .subcommand(GetBalance::build(DisplayOrder::GetBalance as usize))
         .subcommand(GetAccount::build(DisplayOrder::GetAccount as usize))
         .subcommand(GetAuctionInfo::build(DisplayOrder::GetAuctionInfo as usize))
         .subcommand(GetValidatorChanges::build(
@@ -163,7 +160,6 @@ async fn main() {
         QueryGlobalState::NAME => QueryGlobalState::run(matches).await,
         QueryBalance::NAME => QueryBalance::run(matches).await,
         GetDictionaryItem::NAME => GetDictionaryItem::run(matches).await,
-        GetBalance::NAME => GetBalance::run(matches).await,
         GetAccount::NAME => GetAccount::run(matches).await,
         GetAuctionInfo::NAME => GetAuctionInfo::run(matches).await,
         GetValidatorChanges::NAME => GetValidatorChanges::run(matches).await,