+++ src/query_global_state.rs
@@ -3,11 +3,14 @@ use std::{fs, str};
 use async_trait::async_trait;
 use clap::{Arg, ArgGroup, ArgMatches, Command};
 
-use casper_client::cli::{CliError, GlobalStateStrIdentifier, GlobalStateStrParams};
+use casper_client::cli::CliError;
 
 use crate::{command::ClientCommand, common, Success};
 
-const ARG_HEX_STRING: &str = "HEX STRING";
+/// Legacy name of command.
+const COMMAND_ALIAS: &str = "query-state";
+/// Legacy name of block identifier argument.
+const BLOCK_IDENTIFIER_ALIAS: &str = "block-hash";
 
 pub struct QueryGlobalState;
 
@@ -16,56 +19,10 @@ enum DisplayOrder {
     Verbose,
     NodeAddress,
     RpcId,
-    BlockHash,
+    BlockIdentifier,
     StateRootHash,
     Key,
     Path,
-}
-
-mod state_root_hash {
-    use super::*;
-
-    pub(super) const ARG_NAME: &str = "state-root-hash";
-    const ARG_SHORT: char = 's';
-    const ARG_VALUE_NAME: &str = ARG_HEX_STRING;
-    const ARG_HELP: &str = "Hex-encoded hash of the state root";
-
-    pub(super) fn arg() -> Arg<'static> {
-        Arg::new(ARG_NAME)
-            .long(ARG_NAME)
-            .short(ARG_SHORT)
-            .required(false)
-            .value_name(ARG_VALUE_NAME)
-            .help(ARG_HELP)
-            .display_order(DisplayOrder::StateRootHash as usize)
-    }
-
-    pub fn get(matches: &ArgMatches) -> Option<&str> {
-        matches.value_of(ARG_NAME)
-    }
-}
-
-mod block_hash {
-    use super::*;
-
-    pub(super) const ARG_NAME: &str = "block-hash";
-    const ARG_SHORT: char = 'b';
-    const ARG_VALUE_NAME: &str = ARG_HEX_STRING;
-    const ARG_HELP: &str = "Hex-encoded hash of the block";
-
-    pub(super) fn arg() -> Arg<'static> {
-        Arg::new(ARG_NAME)
-            .long(ARG_NAME)
-            .short(ARG_SHORT)
-            .required(false)
-            .value_name(ARG_VALUE_NAME)
-            .help(ARG_HELP)
-            .display_order(DisplayOrder::BlockHash as usize)
-    }
-
-    pub fn get(matches: &ArgMatches) -> Option<&str> {
-        matches.value_of(ARG_NAME)
-    }
 }
 
 /// Handles providing the arg for and retrieval of the key.
@@ -120,7 +77,7 @@ mod key {
                     value, error
                 );
                 CliError::FailedToParsePublicKey {
-                    context: "query",
+                    context: "query".to_string(),
                     error,
                 }
             })?;
@@ -156,34 +113,14 @@ mod path {
     }
 }
 
-fn global_state_str_params(matches: &ArgMatches) -> GlobalStateStrParams<'_> {
-    if let Some(state_root_hash) = state_root_hash::get(matches) {
-        return GlobalStateStrParams {
-            str_identifier: GlobalStateStrIdentifier::Hash {
-                is_block_hash: false,
-            },
-            identifier_value: state_root_hash,
-        };
-    }
-    if let Some(block_hash) = block_hash::get(matches) {
-        return GlobalStateStrParams {
-            str_identifier: GlobalStateStrIdentifier::Hash {
-                is_block_hash: true,
-            },
-            identifier_value: block_hash,
-        };
-    }
-    unreachable!("clap arg groups and parsing should prevent this for global state params")
-}
-
 #[async_trait]
 impl ClientCommand for QueryGlobalState {
     const NAME: &'static str = "query-global-state";
-    const ABOUT: &'static str =
-        "Retrieve a stored value from the network using either the state root hash or block hash";
+    const ABOUT: &'static str = "Retrieve a stored value from the network";
 
     fn build(display_order: usize) -> Command<'static> {
         Command::new(Self::NAME)
+            .alias(COMMAND_ALIAS)
             .about(Self::ABOUT)
             .display_order(display_order)
             .arg(common::verbose::arg(DisplayOrder::Verbose as usize))
@@ -191,23 +128,31 @@ impl ClientCommand for QueryGlobalState {
                 DisplayOrder::NodeAddress as usize,
             ))
             .arg(common::rpc_id::arg(DisplayOrder::RpcId as usize))
-            .arg(key::arg(DisplayOrder::Key as usize))
-            .arg(path::arg(DisplayOrder::Path as usize))
-            .arg(block_hash::arg())
-            .arg(state_root_hash::arg())
+            .arg(
+                common::block_identifier::arg(DisplayOrder::BlockIdentifier as usize, false)
+                    .alias(BLOCK_IDENTIFIER_ALIAS),
+            )
+            .arg(common::state_root_hash::arg(
+                DisplayOrder::StateRootHash as usize,
+                false,
+            ))
             .group(
                 ArgGroup::new("state-identifier")
-                    .arg(state_root_hash::ARG_NAME)
-                    .arg(block_hash::ARG_NAME)
+                    .arg(common::block_identifier::ARG_NAME)
+                    .arg(common::state_root_hash::ARG_NAME)
                     .required(true),
             )
+            .arg(key::arg(DisplayOrder::Key as usize))
+            .arg(path::arg(DisplayOrder::Path as usize))
     }
 
     async fn run(matches: &ArgMatches) -> Result<Success, CliError> {
         let maybe_rpc_id = common::rpc_id::get(matches);
         let node_address = common::node_address::get(matches);
         let verbosity_level = common::verbose::get(matches);
-        let global_state_str_params = global_state_str_params(matches);
+
+        let maybe_block_id = common::block_identifier::get(matches);
+        let maybe_state_root_hash = common::state_root_hash::get(matches).unwrap_or_default();
         let key = key::get(matches)?;
         let path = path::get(matches);
 
@@ -215,7 +160,8 @@ impl ClientCommand for QueryGlobalState {
             maybe_rpc_id,
             node_address,
             verbosity_level,
-            global_state_str_params,
+            maybe_block_id,
+            maybe_state_root_hash,
             &key,
             path,
         )